the final version

+++ y86_new/y86_essence.h
@@ -48,7 +48,7 @@ class M{
 //一下五个是流水线寄存器
 class Freg{
     public:
-    int predPC;
+    long long predPC;
     bool stall;
 }freg;
 class Wreg{
